Renaming, changed collisions, changed controlls
Changed Names of things so that there is a consistent naming convention.
Will write up a doc with the rules for naming stuff later.

Changed the collision code for characters hitting characters, feels
better now but causes a lot of bugs.

Changed the controlls. Aiming and firing are now right and left click
respectively. Aiming also points at the mouse cursor now.

Made a new state machine that uses if-else statements instead of a
switch statement. Allows for more control but looks a bit messier.
Comments will be key for state machines.