Big Player/Enemy Update
Updated statemachines to read stack data structures, that way I can
revert back to the last state I was in. Adds a lot more functionality,
but this does not change the behavior of the bat or fishman enemies.

Have also added a fully working version of the boss, with a preliminary
combat kit.